+++ mobile/android/android-components/components/feature/addons/src/main/java/mozilla/components/feature/addons/AddonsProvider.kt
@@ -10,17 +10,17 @@ package mozilla.components.feature.addons
 interface AddonsProvider {
 
     /**
-     * Provides a list of all available add-ons.
+     * Provides a list of all featured add-ons, which are add-ons we list in the add-ons manager UI
+     * by default (e.g. the add-ons that are available for the app or a set of curated add-ons).
      *
-     * @param allowCache whether or not the result may be provided
-     * from a previously cached response, defaults to true.
-     * @param readTimeoutInSeconds optional timeout in seconds to use when fetching
-     * available add-ons from a remote endpoint.
+     * @param allowCache whether or not the result may be provided from a previously cached response,
+     * defaults to true.
+     * @param readTimeoutInSeconds optional timeout in seconds to use when fetching featured add-ons.
      * @param language indicates in which language the translatable fields should be in, if no
      * matching language is found then a fallback translation is returned using the default language.
      * When it is null all translations available will be returned.
      */
-    suspend fun getAvailableAddons(
+    suspend fun getFeaturedAddons(
         allowCache: Boolean = true,
         readTimeoutInSeconds: Long? = null,
         language: String? = null,