+++ mobile/android/android-components/components/feature/addons/src/test/java/mozilla/components/feature/addons/amo/AMOAddonsProviderTest.kt
@@ -34,13 +34,13 @@ import java.util.Date
 import java.util.concurrent.TimeUnit
 
 @RunWith(AndroidJUnit4::class)
-class AddonCollectionProviderTest {
+class AMOAddonsProviderTest {
 
     @Test
-    fun `getAvailableAddons - with a successful status response must contain add-ons`() = runTest {
+    fun `getFeaturedAddons - with a successful status response must contain add-ons`() = runTest {
         val mockedClient = prepareClient(loadResourceAsString("/collection.json"))
-        val provider = AddonCollectionProvider(testContext, client = mockedClient)
-        val addons = provider.getAvailableAddons()
+        val provider = AMOAddonsProvider(testContext, client = mockedClient)
+        val addons = provider.getFeaturedAddons()
         val addon = addons.first()
 
         assertTrue(addons.isNotEmpty())
@@ -96,11 +96,11 @@ class AddonCollectionProviderTest {
     }
 
     @Test
-    fun `getAvailableAddons - with a successful status response must handle empty values`() = runTest {
+    fun `getFeaturedAddons - with a successful status response must handle empty values`() = runTest {
         val client = prepareClient()
-        val provider = AddonCollectionProvider(testContext, client = client)
+        val provider = AMOAddonsProvider(testContext, client = client)
 
-        val addons = provider.getAvailableAddons()
+        val addons = provider.getFeaturedAddons()
         val addon = addons.first()
 
         assertTrue(addons.isNotEmpty())
@@ -135,11 +135,11 @@ class AddonCollectionProviderTest {
     }
 
     @Test
-    fun `getAvailableAddons - with a language`() = runTest {
+    fun `getFeaturedAddons - with a language`() = runTest {
         val client = prepareClient(loadResourceAsString("/localized_collection.json"))
-        val provider = AddonCollectionProvider(testContext, client = client)
+        val provider = AMOAddonsProvider(testContext, client = client)
 
-        val addons = provider.getAvailableAddons(language = "en")
+        val addons = provider.getFeaturedAddons(language = "en")
         val addon = addons.first()
 
         assertTrue(addons.isNotEmpty())
@@ -208,11 +208,11 @@ class AddonCollectionProviderTest {
     }
 
     @Test
-    fun `getAvailableAddons - read timeout can be configured`() = runTest {
+    fun `getFeaturedAddons - read timeout can be configured`() = runTest {
         val mockedClient = prepareClient()
 
-        val provider = spy(AddonCollectionProvider(testContext, client = mockedClient))
-        provider.getAvailableAddons(readTimeoutInSeconds = 5)
+        val provider = spy(AMOAddonsProvider(testContext, client = mockedClient))
+        provider.getFeaturedAddons(readTimeoutInSeconds = 5)
         verify(mockedClient).fetch(
             Request(
                 url = "https://services.addons.mozilla.org/api/v4/accounts/account/mozilla/collections/" +
@@ -224,43 +224,43 @@ class AddonCollectionProviderTest {
     }
 
     @Test(expected = IOException::class)
-    fun `getAvailableAddons - with unexpected status will throw exception`() = runTest {
+    fun `getFeaturedAddons - with unexpected status will throw exception`() = runTest {
         val mockedClient = prepareClient(status = 500)
-        val provider = AddonCollectionProvider(testContext, client = mockedClient)
-        provider.getAvailableAddons()
+        val provider = AMOAddonsProvider(testContext, client = mockedClient)
+        provider.getFeaturedAddons()
         Unit
     }
 
     @Test
-    fun `getAvailableAddons - returns cached result if allowed and not expired`() = runTest {
+    fun `getFeaturedAddons - returns cached result if allowed and not expired`() = runTest {
         val mockedClient = prepareClient(loadResourceAsString("/collection.json"))
 
-        val provider = spy(AddonCollectionProvider(testContext, client = mockedClient))
-        provider.getAvailableAddons(false)
+        val provider = spy(AMOAddonsProvider(testContext, client = mockedClient))
+        provider.getFeaturedAddons(false)
         verify(provider, never()).readFromDiskCache(null, useFallbackFile = false)
 
         whenever(provider.cacheExpired(testContext, null, useFallbackFile = false)).thenReturn(true)
-        provider.getAvailableAddons(true)
+        provider.getFeaturedAddons(true)
         verify(provider, never()).readFromDiskCache(null, useFallbackFile = false)
 
         whenever(provider.cacheExpired(testContext, null, useFallbackFile = false)).thenReturn(false)
-        provider.getAvailableAddons(true)
+        provider.getFeaturedAddons(true)
         verify(provider).readFromDiskCache(null, useFallbackFile = false)
         Unit
     }
 
     @Test
-    fun `getAvailableAddons - returns cached result if allowed and fetch failed`() = runTest {
+    fun `getFeaturedAddons - returns cached result if allowed and fetch failed`() = runTest {
         val mockedClient: Client = mock()
         val exception = IOException("test")
         val cachedAddons: List<Addon> = emptyList()
         whenever(mockedClient.fetch(any())).thenThrow(exception)
 
-        val provider = spy(AddonCollectionProvider(testContext, client = mockedClient))
+        val provider = spy(AMOAddonsProvider(testContext, client = mockedClient))
 
         try {
             // allowCache = false
-            provider.getAvailableAddons(allowCache = false)
+            provider.getFeaturedAddons(allowCache = false)
             fail("Expected IOException")
         } catch (e: IOException) {
             assertSame(exception, e)
@@ -268,7 +268,7 @@ class AddonCollectionProviderTest {
 
         try {
             // allowCache = true, but no cache present
-            provider.getAvailableAddons(allowCache = true)
+            provider.getFeaturedAddons(allowCache = true)
             fail("Expected IOException")
         } catch (e: IOException) {
             assertSame(exception, e)
@@ -277,7 +277,7 @@ class AddonCollectionProviderTest {
         try {
             // allowCache = true, cache present, but we fail to read
             whenever(provider.getCacheLastUpdated(testContext, null, useFallbackFile = false)).thenReturn(Date().time)
-            provider.getAvailableAddons(allowCache = true)
+            provider.getFeaturedAddons(allowCache = true)
             fail("Expected IOException")
         } catch (e: IOException) {
             assertSame(exception, e)
@@ -286,38 +286,38 @@ class AddonCollectionProviderTest {
 //        // allowCache = true, cache present for a fallback file, and reading successfully
         whenever(provider.getCacheLastUpdated(testContext, null, useFallbackFile = true)).thenReturn(Date().time)
         whenever(provider.readFromDiskCache(null, useFallbackFile = true)).thenReturn(cachedAddons)
-        assertSame(cachedAddons, provider.getAvailableAddons(allowCache = true))
+        assertSame(cachedAddons, provider.getFeaturedAddons(allowCache = true))
 
         // allowCache = true, cache present, and reading successfully
         whenever(provider.getCacheLastUpdated(testContext, null, useFallbackFile = false)).thenReturn(Date().time)
         whenever(provider.cacheExpired(testContext, null, useFallbackFile = false)).thenReturn(false)
         whenever(provider.readFromDiskCache(null, useFallbackFile = false)).thenReturn(cachedAddons)
-        assertSame(cachedAddons, provider.getAvailableAddons(allowCache = true))
+        assertSame(cachedAddons, provider.getFeaturedAddons(allowCache = true))
     }
 
     @Test
-    fun `getAvailableAddons - writes response to cache if configured`() = runTest {
+    fun `getFeaturedAddons - writes response to cache if configured`() = runTest {
         val jsonResponse = loadResourceAsString("/collection.json")
         val mockedClient = prepareClient(jsonResponse)
 
-        val provider = spy(AddonCollectionProvider(testContext, client = mockedClient))
-        val cachingProvider = spy(AddonCollectionProvider(testContext, client = mockedClient, maxCacheAgeInMinutes = 1))
+        val provider = spy(AMOAddonsProvider(testContext, client = mockedClient))
+        val cachingProvider = spy(AMOAddonsProvider(testContext, client = mockedClient, maxCacheAgeInMinutes = 1))
 
-        provider.getAvailableAddons()
+        provider.getFeaturedAddons()
         verify(provider, never()).writeToDiskCache(jsonResponse, null)
 
-        cachingProvider.getAvailableAddons()
+        cachingProvider.getFeaturedAddons()
         verify(cachingProvider).writeToDiskCache(jsonResponse, null)
     }
 
     @Test
-    fun `getAvailableAddons - deletes unused cache files`() = runTest {
+    fun `getFeaturedAddons - deletes unused cache files`() = runTest {
         val jsonResponse = loadResourceAsString("/collection.json")
         val mockedClient = prepareClient(jsonResponse)
 
-        val provider = spy(AddonCollectionProvider(testContext, client = mockedClient, maxCacheAgeInMinutes = 1))
+        val provider = spy(AMOAddonsProvider(testContext, client = mockedClient, maxCacheAgeInMinutes = 1))
 
-        provider.getAvailableAddons()
+        provider.getFeaturedAddons()
         verify(provider).deleteUnusedCacheFiles(null)
     }
 
@@ -335,7 +335,7 @@ class AddonCollectionProviderTest {
         collectionFile.createNewFile()
         assertTrue(collectionFile.exists())
 
-        val provider = AddonCollectionProvider(testContext, client = prepareClient(), maxCacheAgeInMinutes = 1)
+        val provider = AMOAddonsProvider(testContext, client = prepareClient(), maxCacheAgeInMinutes = 1)
         provider.deleteUnusedCacheFiles(null)
         assertTrue(regularFile.exists())
         assertTrue(regularDir.exists())
@@ -352,7 +352,7 @@ class AddonCollectionProviderTest {
         regularDir.mkdir()
         assertTrue(regularDir.exists())
 
-        val provider = AddonCollectionProvider(testContext, client = prepareClient(), maxCacheAgeInMinutes = 1)
+        val provider = AMOAddonsProvider(testContext, client = prepareClient(), maxCacheAgeInMinutes = 1)
         val enFile = File(testContext.filesDir, provider.getCacheFileName("en"))
 
         enFile.createNewFile()
@@ -376,7 +376,7 @@ class AddonCollectionProviderTest {
 
     @Test
     fun `getBaseCacheFile - will return a first localized file WHEN the provided language file is not available`() {
-        val provider = AddonCollectionProvider(testContext, client = prepareClient(), maxCacheAgeInMinutes = 1)
+        val provider = AMOAddonsProvider(testContext, client = prepareClient(), maxCacheAgeInMinutes = 1)
         val enFile = File(testContext.filesDir, provider.getCacheFileName("en"))
 
         enFile.createNewFile()
@@ -394,15 +394,15 @@ class AddonCollectionProviderTest {
     }
 
     @Test
-    fun `getAvailableAddons - cache expiration check`() {
-        var provider = spy(AddonCollectionProvider(testContext, client = mock(), maxCacheAgeInMinutes = -1))
+    fun `getFeaturedAddons - cache expiration check`() {
+        var provider = spy(AMOAddonsProvider(testContext, client = mock(), maxCacheAgeInMinutes = -1))
         whenever(provider.getCacheLastUpdated(testContext, null, useFallbackFile = false)).thenReturn(Date().time)
         assertTrue(provider.cacheExpired(testContext, null, useFallbackFile = false))
 
         whenever(provider.getCacheLastUpdated(testContext, null, useFallbackFile = false)).thenReturn(-1)
         assertTrue(provider.cacheExpired(testContext, null, useFallbackFile = false))
 
-        provider = spy(AddonCollectionProvider(testContext, client = mock(), maxCacheAgeInMinutes = 10))
+        provider = spy(AMOAddonsProvider(testContext, client = mock(), maxCacheAgeInMinutes = 10))
         whenever(provider.getCacheLastUpdated(testContext, null, useFallbackFile = false)).thenReturn(-1)
         assertTrue(provider.cacheExpired(testContext, null, useFallbackFile = false))
 
@@ -424,7 +424,7 @@ class AddonCollectionProviderTest {
         whenever(mockedResponse.status).thenReturn(200)
         whenever(mockedClient.fetch(any())).thenReturn(mockedResponse)
 
-        val provider = AddonCollectionProvider(testContext, client = mockedClient)
+        val provider = AMOAddonsProvider(testContext, client = mockedClient)
         val addon = Addon(
             id = "id",
             authors = mock(),
@@ -443,7 +443,7 @@ class AddonCollectionProviderTest {
     @Test
     fun `getAddonIconBitmap - with an unsuccessful status will return null`() = runTest {
         val mockedClient = prepareClient(status = 500)
-        val provider = AddonCollectionProvider(testContext, client = mockedClient)
+        val provider = AMOAddonsProvider(testContext, client = mockedClient)
         val addon = Addon(
             id = "id",
             authors = mock(),
@@ -464,13 +464,13 @@ class AddonCollectionProviderTest {
         val mockedClient = prepareClient()
 
         val collectionName = "collection123"
-        val provider = AddonCollectionProvider(
+        val provider = AMOAddonsProvider(
             testContext,
             client = mockedClient,
             collectionName = collectionName,
         )
 
-        provider.getAvailableAddons()
+        provider.getFeaturedAddons()
         verify(mockedClient).fetch(
             Request(
                 url = "https://services.addons.mozilla.org/api/v4/accounts/account/mozilla/collections/" +
@@ -487,13 +487,13 @@ class AddonCollectionProviderTest {
         val mockedClient = prepareClient()
 
         val collectionName = "collection123"
-        AddonCollectionProvider(
+        AMOAddonsProvider(
             testContext,
             client = mockedClient,
             collectionName = collectionName,
             sortOption = SortOption.POPULARITY,
         ).also {
-            it.getAvailableAddons()
+            it.getFeaturedAddons()
         }
 
         verify(mockedClient).fetch(
@@ -504,13 +504,13 @@ class AddonCollectionProviderTest {
             ),
         )
 
-        AddonCollectionProvider(
+        AMOAddonsProvider(
             testContext,
             client = mockedClient,
             collectionName = collectionName,
             sortOption = SortOption.POPULARITY_DESC,
         ).also {
-            it.getAvailableAddons()
+            it.getFeaturedAddons()
         }
 
         verify(mockedClient).fetch(
@@ -521,13 +521,13 @@ class AddonCollectionProviderTest {
             ),
         )
 
-        AddonCollectionProvider(
+        AMOAddonsProvider(
             testContext,
             client = mockedClient,
             collectionName = collectionName,
             sortOption = SortOption.NAME,
         ).also {
-            it.getAvailableAddons()
+            it.getFeaturedAddons()
         }
 
         verify(mockedClient).fetch(
@@ -538,13 +538,13 @@ class AddonCollectionProviderTest {
             ),
         )
 
-        AddonCollectionProvider(
+        AMOAddonsProvider(
             testContext,
             client = mockedClient,
             collectionName = collectionName,
             sortOption = SortOption.NAME_DESC,
         ).also {
-            it.getAvailableAddons()
+            it.getFeaturedAddons()
         }
 
         verify(mockedClient).fetch(
@@ -555,13 +555,13 @@ class AddonCollectionProviderTest {
             ),
         )
 
-        AddonCollectionProvider(
+        AMOAddonsProvider(
             testContext,
             client = mockedClient,
             collectionName = collectionName,
             sortOption = SortOption.DATE_ADDED,
         ).also {
-            it.getAvailableAddons()
+            it.getFeaturedAddons()
         }
 
         verify(mockedClient).fetch(
@@ -572,13 +572,13 @@ class AddonCollectionProviderTest {
             ),
         )
 
-        AddonCollectionProvider(
+        AMOAddonsProvider(
             testContext,
             client = mockedClient,
             collectionName = collectionName,
             sortOption = SortOption.DATE_ADDED_DESC,
         ).also {
-            it.getAvailableAddons()
+            it.getFeaturedAddons()
         }
 
         verify(mockedClient).fetch(
@@ -597,14 +597,14 @@ class AddonCollectionProviderTest {
         val mockedClient = prepareClient()
         val collectionUser = "user123"
         val collectionName = "collection123"
-        val provider = AddonCollectionProvider(
+        val provider = AMOAddonsProvider(
             testContext,
             client = mockedClient,
             collectionUser = collectionUser,
             collectionName = collectionName,
         )
 
-        provider.getAvailableAddons()
+        provider.getFeaturedAddons()
         verify(mockedClient).fetch(
             Request(
                 url = "https://services.addons.mozilla.org/api/v4/accounts/account/" +
@@ -625,12 +625,12 @@ class AddonCollectionProviderTest {
     fun `default collection is used if not configured`() = runTest {
         val mockedClient = prepareClient()
 
-        val provider = AddonCollectionProvider(
+        val provider = AMOAddonsProvider(
             testContext,
             client = mockedClient,
         )
 
-        provider.getAvailableAddons()
+        provider.getFeaturedAddons()
         verify(mockedClient).fetch(
             Request(
                 url = "https://services.addons.mozilla.org/api/v4/accounts/account/" +
@@ -649,7 +649,7 @@ class AddonCollectionProviderTest {
         val mockedClient = prepareClient()
         val collectionUser = "../../user"
         val collectionName = "../collection"
-        val provider = AddonCollectionProvider(
+        val provider = AMOAddonsProvider(
             testContext,
             client = mockedClient,
             collectionUser = collectionUser,