+++ mobile/android/android-components/components/feature/addons/src/test/java/mozilla/components/feature/addons/ui/AddonInstallationDialogFragmentTest.kt
@@ -18,7 +18,7 @@ import kotlinx.coroutines.Job
 import kotlinx.coroutines.test.runTest
 import mozilla.components.feature.addons.Addon
 import mozilla.components.feature.addons.R
-import mozilla.components.feature.addons.amo.AddonCollectionProvider
+import mozilla.components.feature.addons.amo.AMOAddonsProvider
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.robolectric.testContext
 import mozilla.components.support.test.rule.MainCoroutineRule
@@ -55,9 +55,9 @@ class AddonInstallationDialogFragmentTest {
             translatableName = mapOf(Addon.DEFAULT_LOCALE to "my_addon"),
             permissions = listOf("privacy", "<all_urls>", "tabs"),
         )
-        val mockedCollectionProvider = mock<AddonCollectionProvider>()
+        val mockedCollectionProvider = mock<AMOAddonsProvider>()
         val fragment = createAddonInstallationDialogFragment(addon, mockedCollectionProvider)
-        assertSame(mockedCollectionProvider, fragment.addonCollectionProvider)
+        assertSame(mockedCollectionProvider, fragment.addonsProvider)
         assertSame(addon, fragment.arguments?.getParcelableCompat(KEY_INSTALLED_ADDON, Addon::class.java))
 
         doReturn(testContext).`when`(fragment).requireContext()
@@ -162,7 +162,7 @@ class AddonInstallationDialogFragmentTest {
         val addon = mock<Addon>()
         val bitmap = mock<Bitmap>()
         val mockedImageView = spy(ImageView(testContext))
-        val mockedCollectionProvider = mock<AddonCollectionProvider>()
+        val mockedCollectionProvider = mock<AMOAddonsProvider>()
         val fragment = createAddonInstallationDialogFragment(addon, mockedCollectionProvider)
 
         whenever(mockedCollectionProvider.getAddonIconBitmap(addon)).thenReturn(bitmap)
@@ -176,7 +176,7 @@ class AddonInstallationDialogFragmentTest {
     fun `handle errors while fetching the add-on icon`() = runTest {
         val addon = mock<Addon>()
         val mockedImageView = spy(ImageView(testContext))
-        val mockedCollectionProvider = mock<AddonCollectionProvider>()
+        val mockedCollectionProvider = mock<AMOAddonsProvider>()
         val fragment = createAddonInstallationDialogFragment(addon, mockedCollectionProvider)
 
         whenever(mockedCollectionProvider.getAddonIconBitmap(addon)).then {
@@ -193,7 +193,7 @@ class AddonInstallationDialogFragmentTest {
     @Test
     fun `allows state loss when committing`() {
         val addon = mock<Addon>()
-        val mockedCollectionProvider = mock<AddonCollectionProvider>()
+        val mockedCollectionProvider = mock<AMOAddonsProvider>()
         val fragment = createAddonInstallationDialogFragment(addon, mockedCollectionProvider)
 
         val fragmentManager = mock<FragmentManager>()
@@ -207,7 +207,7 @@ class AddonInstallationDialogFragmentTest {
     @Test
     fun `cancels icon job on stop`() {
         val addon = mock<Addon>()
-        val mockedCollectionProvider = mock<AddonCollectionProvider>()
+        val mockedCollectionProvider = mock<AMOAddonsProvider>()
         val fragment = createAddonInstallationDialogFragment(addon, mockedCollectionProvider)
 
         val job = mock<Job>()
@@ -218,10 +218,10 @@ class AddonInstallationDialogFragmentTest {
 
     private fun createAddonInstallationDialogFragment(
         addon: Addon,
-        addonCollectionProvider: AddonCollectionProvider,
+        addonsProvider: AMOAddonsProvider,
         promptsStyling: AddonInstallationDialogFragment.PromptsStyling? = null,
     ): AddonInstallationDialogFragment {
-        return spy(AddonInstallationDialogFragment.newInstance(addon, addonCollectionProvider, promptsStyling = promptsStyling)).apply {
+        return spy(AddonInstallationDialogFragment.newInstance(addon, addonsProvider, promptsStyling = promptsStyling)).apply {
             doNothing().`when`(this).dismiss()
         }
     }