+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/components/Components.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.core.app.NotificationManagerCompat
 import com.google.android.play.core.review.ReviewManagerFactory
 import mozilla.components.feature.addons.AddonManager
-import mozilla.components.feature.addons.amo.AddonCollectionProvider
+import mozilla.components.feature.addons.amo.AMOAddonsProvider
 import mozilla.components.feature.addons.migration.DefaultSupportedAddonsChecker
 import mozilla.components.feature.addons.update.DefaultAddonUpdater
 import mozilla.components.feature.autofill.AutofillConfiguration
@@ -112,10 +112,10 @@ class Components(private val context: Context) {
         )
     }
 
-    val addonCollectionProvider by lazyMonitored {
+    val addonsProvider by lazyMonitored {
         // Check if we have a customized (overridden) AMO collection (supported in Nightly & Beta)
         if (FeatureFlags.customExtensionCollectionFeature && context.settings().amoCollectionOverrideConfigured()) {
-            AddonCollectionProvider(
+            AMOAddonsProvider(
                 context,
                 core.client,
                 collectionUser = context.settings().overrideAmoUser,
@@ -126,7 +126,7 @@ class Components(private val context: Context) {
         else if (!BuildConfig.AMO_COLLECTION_USER.isNullOrEmpty() &&
             !BuildConfig.AMO_COLLECTION_NAME.isNullOrEmpty()
         ) {
-            AddonCollectionProvider(
+            AMOAddonsProvider(
                 context,
                 core.client,
                 serverURL = BuildConfig.AMO_SERVER_URL,
@@ -137,7 +137,7 @@ class Components(private val context: Context) {
         }
         // Fall back to defaults
         else {
-            AddonCollectionProvider(context, core.client, maxCacheAgeInMinutes = AMO_COLLECTION_MAX_CACHE_AGE)
+            AMOAddonsProvider(context, core.client, maxCacheAgeInMinutes = AMO_COLLECTION_MAX_CACHE_AGE)
         }
     }
 
@@ -155,7 +155,7 @@ class Components(private val context: Context) {
     }
 
     val addonManager by lazyMonitored {
-        AddonManager(core.store, core.engine, addonCollectionProvider, addonUpdater)
+        AddonManager(core.store, core.engine, addonsProvider, addonUpdater)
     }
 
     val analytics by lazyMonitored { Analytics(context) }